+++ cyber/logger/logger.cc
@@ -43,7 +43,7 @@ Logger::~Logger() {
 }
 
 void Logger::Write(bool force_flush, time_t timestamp, const char* message,
-                   int message_len) {
+                   size_t message_len) {
   std::string log_message = std::string(message, message_len);
   std::string module_name;
   // set the same bracket as the bracket in log.h
