+++ cyber/logger/logger.h
@@ -30,7 +30,7 @@ class Logger : public google::base::Logger {
   explicit Logger(google::base::Logger* wrapped);
   ~Logger();
   void Write(bool force_flush, time_t timestamp, const char* message,
-             int message_len) override;
+             size_t message_len) override;
   void Flush() override;
   uint32_t LogSize() override;
